Move `CreateDelegationSession` into its own file
